+++ apps/api/src/main/java/com/fridgetoplate/service/ReviewService.java
@@ -1,0 +1,26 @@
+package com.fridgetoplate.service;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
+import com.fridgetoplate.model.Review;
+import com.fridgetoplate.repository.ReviewRepository;
+
+public class ReviewService {
+    
+
+    @Autowired 
+    private ReviewRepository reviewRepository;
+
+     public String removeReviews(List<Review> reviews) {
+        reviewRepository.removeReviews(reviews);
+        return "REVIEWS SUCCESSFULLY DELETED";
+    }
+
+    public List<Review> getReviewsById(String recipeId) {
+        return reviewRepository.getReviewsById(recipeId);   
+    }
+
+
+}
